Pubdate applied to Archives. Midnight date check added. JSON Update
